Adding BST/BSR Map Heuristic and Game Team Class
Made some small changes to GameMap and created a new class for GameTeam.
I figure it will probably be best to have multiple classes because
sometimes you want to work with inividual teams and sometimes you want
to consider the whole map. The one downside of this is making sure that
you update everything properly.

So... if team 'red' takes Yakutsk, you must update the game map and the
game team variables so you can properly reference both in the future.
We could create a function for this.

Made some changes to the file structure as well with putting utils in a
subdirectory.