Complete Game Simulation Setup
With the last commit, I have made changes to four different files in
order to allow an automated strategy to play the game for you.
Essentially, the policy returns a move, and the game will simulate
the dice rolls for that move until someone wins/loses that battle.
Then, it picks another move. There is some forced code in this
but it allows a bit of flexibility. I will go through all of my changes
one by one.

strategy.py (Strategy Class): This creates a Strategy Superclass that
an AI can folow. The children of the Strategy class (also on strategy.py)
inherit the same qualities. For example, I have created the RandomStrategy
subclass of the StrategyClass with two characteristics: addTroopsTurn and
getNextMove. This will add troops and pick moves at random. As you will be
able to see, each strategy is associated with a game_team (after all, a
policy for one team will return different attacks than a policy for another
team)

game_team_class.py (GameTeam Class): I added onto the GameTeam class in
order to support the abilities of playing and having a strategy. I also
added plenty of comments here. Within the changes is a new attribute for
the strategy (which references an instance of the Strategy Class). From
there, I created several different functions that allow you to manually
make moves or to reference the strategy to make a move for you. This
could allow us (as humans) to play against the RISK AI agent if we wanted!
Lastly, I made a function for playTurn() which references the strategy
attribute again. This tells the strategy to execute moves according to its
policy for a turn. For each of the moves, it calls "makeMove" which calls
simNextMove() which actually plays out the attack until completion (NOTE:
there is no stop condition here). Then, the map and the team classes will
update respectively!

map_setup_functions.py: This is another reference Python script with relevant
scripts for setting up the map. initializeFullRiskMap() has the large
declarations that Jane wrote for the full Risk Map and consolidates it. It
also contains setGameBoardRandom() which initializes a random board game
states (creates teams, fills map, etc.) for a list of teams (can be more than
one)

simulate_random_game.ipynb (Jupyter Notebook): This is where I am developing
the start of a game with two random players. This is not completed yet but
demonstrates how simple the new representation can be for creating a new game!