Lookahead w/ Rollouts Partially Complete
I ran 100 simulations w/ L.A.R against Rule of Thumb and took
the statistics. The results are stored in a .npy file for the
importan information from those simulations. The rest of the
changes were small ones in the .py files in order to faciliate
the lookahead properly. Currently working with a lookahead of
depth 4 for the data I took.